# [WIP]: Part 3 (FINAL): Improve performance of PerWikiCourseStats#stats