@Abishekcs 2025-08-14
Same as https://github.com/WikiEducationFoundation/WikiEduDashboard/pull/6438#issuecomment-3189168551